spdylay_ssl.cc: Cast tv.tv_sec to long int to pass printf

+++ src/spdylay_ssl.cc
@@ -516,7 +516,7 @@ void print_timer()
   get_timer(&tv);
   printf("%s[%3ld.%03ld]%s",
          ansi_esc("\033[33m"),
-         tv.tv_sec, tv.tv_usec/1000,
+         (long int)tv.tv_sec, tv.tv_usec/1000,
          ansi_escend());
 }
